+++ apps/platform/src/hooks/useUserStore.ts
@@ -4,8 +4,12 @@ import { createJSONStorage, persist } from 'zustand/middleware'
 export interface UserStore {
   /** Show disclaimer if first time user */
   newcomer: boolean
+  /** User Modal */
+  userModal: boolean
   /** A list of games played. The first time a game is opened we can display info */
   gamesPlayed: Array<string>
+  /** The last pool a user had selected */
+  lastSelectedPool: { token: string, authority?: string } | null
   markGameAsPlayed: (gameId: string, played: boolean) => void
   set: StoreApi<UserStore>['setState']
 }
@@ -17,6 +21,8 @@ export const useUserStore = create(
   persist<UserStore>(
     (set, get) => ({
       newcomer: true,
+      userModal: false,
+      lastSelectedPool: null,
       gamesPlayed: [],
       markGameAsPlayed: (gameId, played) => {
         const gamesPlayed = new Set(get().gamesPlayed)
@@ -31,7 +37,7 @@ export const useUserStore = create(
     }),
     {
       name: 'user',
-      storage: createJSONStorage(() => sessionStorage),
+      storage: createJSONStorage(() => window.localStorage),
     },
   ),
 )