+++ apps/platform/src/sections/Game/Game.tsx
@@ -1,4 +1,4 @@
-import { GambaUi, useGambaAudioStore } from 'gamba-react-ui-v2'
+import { GambaUi, useSoundStore } from 'gamba-react-ui-v2'
 import React from 'react'
 import { useParams } from 'react-router-dom'
 import { Icon } from '../../components/Icon'
@@ -6,10 +6,10 @@ import { Modal } from '../../components/Modal'
 import { GAMES } from '../../games'
 import { useUserStore } from '../../hooks/useUserStore'
 import { GameSlider } from '../Dashboard/Dashboard'
-import { Container, Controls, IconButton, MetaControls, Screen, Splash } from './Game.styles'
-import { LoadingBar } from './LoadingBar'
+import { Container, Controls, IconButton, MetaControls, Screen, Spinner, Splash } from './Game.styles'
+import { LoadingBar, useLoadingState } from './LoadingBar'
 import { ProvablyFairModal } from './ProvablyFairModal'
-// import { TransactionModal } from './TransactionModal'
+import { TransactionModal } from './TransactionModal'
 
 function CustomError() {
   return (
@@ -33,11 +33,11 @@ function CustomRenderer() {
   const { game } = GambaUi.useGame()
   const [info, setInfo] = React.useState(false)
   const [provablyFair, setProvablyFair] = React.useState(false)
-  // const [txModal, setTransactionModal] = React.useState(false)
-  const audioStore = useGambaAudioStore()
+  const soundStore = useSoundStore()
   const firstTimePlaying = useUserStore((state) => !state.gamesPlayed.includes(game.id))
   const markGameAsPlayed = useUserStore((state) => () => state.markGameAsPlayed(game.id, true))
   const [ready, setReady] = React.useState(false)
+  const [txModal, setTxModal] = React.useState(false)
   // const loading = useLoadingState()
 
   React.useEffect(
@@ -68,7 +68,7 @@ function CustomRenderer() {
   return (
     <>
       {info && (
-        <Modal onClose={() => setInfo(false)}>
+        <Modal onClose={closeInfo}>
           <h1>
             <img height="100px" title={game.meta.name} src={game.meta.image} />
           </h1>
@@ -81,9 +81,9 @@ function CustomRenderer() {
       {provablyFair && (
         <ProvablyFairModal onClose={() => setProvablyFair(false)} />
       )}
-      {/* {txModal && (
-        <TransactionModal onClose={() => setTransactionModal(false)} />
-      )} */}
+      {txModal && (
+        <TransactionModal onClose={() => setTxModal(false)} />
+      )}
       <Container>
         <Screen>
           <Splash>
@@ -92,22 +92,8 @@ function CustomRenderer() {
           <GambaUi.PortalTarget target="error" />
           {ready && <GambaUi.PortalTarget target="screen" />}
           <MetaControls>
-            <IconButton onClick={() => setInfo(true)}>
-              <Icon.Info />
-            </IconButton>
-            <IconButton onClick={() => setProvablyFair(true)}>
-              <Icon.Fairness />
-            </IconButton>
-            <IconButton onClick={() => audioStore.set(audioStore.masterGain ? 0 : .5)}>
-              {audioStore.masterGain ? '🔈' : '🔇'}
-            </IconButton>
-          </MetaControls>
-        </Screen>
-        <LoadingBar />
-        <Controls>
-          <div style={{ display: 'flex', gap: '10px' }}>
             {/* <div style={{ display: 'flex' }}>
-              <IconButton onClick={() => setTransactionModal(true)}>
+              <IconButton onClick={() => setTxModal(true)}>
                 {loading === -1 ? (
                   <Icon.Shuffle />
                 ) : (
@@ -115,10 +101,22 @@ function CustomRenderer() {
                 )}
               </IconButton>
             </div> */}
-            <div style={{ display: 'flex', alignItems: 'center', gap: '10px' }}>
-              <GambaUi.PortalTarget target="controls" />
-              <GambaUi.PortalTarget target="play" />
-            </div>
+            <IconButton onClick={() => setInfo(true)}>
+              <Icon.Info />
+            </IconButton>
+            <IconButton onClick={() => setProvablyFair(true)}>
+              <Icon.Fairness />
+            </IconButton>
+            <IconButton onClick={() => soundStore.set(soundStore.volume ? 0 : .5)}>
+              {soundStore.volume ? <Icon.Volume /> : <Icon.VolumeMuted />}
+            </IconButton>
+          </MetaControls>
+        </Screen>
+        <LoadingBar />
+        <Controls>
+          <div style={{ display: 'flex', alignItems: 'center', gap: '10px' }}>
+            <GambaUi.PortalTarget target="controls" />
+            <GambaUi.PortalTarget target="play" />
           </div>
         </Controls>
       </Container>