+++ apps/website/components/projects.ts
@@ -1,0 +1,37 @@
+export const projects = [
+  {
+    name: 'Guacamole GG',
+    link: 'https://www.guacamole.gg/play',
+    thumbnail: '/thumbnails/guac.png',
+  },
+  {
+    name: 'Fabled Fortune',
+    link: 'https://fabled-fortune.fabled-adventure.xyz',
+    thumbnail: '/thumbnails/fabled-fortune.png',
+  },
+  {
+    name: 'Oh Yeah Bet',
+    link: 'https://www.ohyeah.bet/',
+    thumbnail: '/thumbnails/ohyeah.png',
+  },
+  {
+    name: 'Fronk Casino',
+    link: 'https://casino.fronk.xyz/',
+    thumbnail: '/thumbnails/fronk.png',
+  },
+  {
+    name: 'Sol Bets',
+    link: 'https://www.solbets.app/',
+    thumbnail: '/thumbnails/solbets.png',
+  },
+  {
+    name: 'Felt Gaming Casino',
+    link: 'https://casino.feltgaming.io/',
+    thumbnail: '/thumbnails/feltgaming.png',
+  },
+  {
+    name: 'NinjaTSS Casino',
+    link: 'https://www.ninjatsscasino.xyz/',
+    thumbnail: '/thumbnails/ninjaturtles.png',
+  },
+]