+++ packages/react-ui/src/components/Select.tsx
@@ -8,13 +8,14 @@ export interface SelectProps<T> extends React.PropsWithChildren {
   options: T[]
   onChange: (value: T) => void
   label?: (value: T) => React.ReactNode
+  className?: string
 }
 
 const StyledWrapper = styled.div`
   position: relative;
 `
 
-export const StyledPopup = styled.div`
+const StyledPopup = styled.div`
   position: absolute;
   bottom: 100%;
   left: 0;
@@ -52,7 +53,7 @@ export function Select<T>(props: SelectProps<T>) {
   }
 
   return (
-    <StyledWrapper>
+    <StyledWrapper className={props.className}>
       <Button disabled={props.disabled} onClick={() => setOpen(!open)}>
         {props.label ? props.label(props.value) : JSON.stringify(props.value)}
       </Button>