+++ packages/react-ui/src/components/Switch.tsx
@@ -9,6 +9,9 @@ const SwitchButton = styled.input`
   aspect-ratio: 2 / 1;
   border-radius: 20px;
   background: #0e0e16;
+  &:checked {
+    background: #9564ff66;
+  }
   &:checked:after {
     left: 50%;
   }
@@ -27,10 +30,8 @@ const SwitchButton = styled.input`
     outline: #9564ff solid 1px;
     outline-offset: 1px;
   }
-  &:disabled {
-    &:after {
-      background: gray;
-    }
+  &:disabled:after {
+    background: gray;
   }
 `
 
@@ -38,12 +39,14 @@ interface SwitchProps {
   checked?: boolean
   onChange?: (checked: boolean) => void
   disabled?: boolean
+  className?: string
 }
 
 export function Switch(props: SwitchProps) {
   return (
     <SwitchButton
       type="checkbox"
+      className={props.className}
       checked={props.checked}
       disabled={props.disabled}
       onChange={(evt) => props.onChange && props.onChange(evt.target.checked)}