+++ packages/react-ui/CHANGELOG.md
@@ -1,5 +1,47 @@
 # gamba-react-ui-v2
 
+## 0.7.0
+
+### Minor Changes
+
+- Referrals
+
+## 0.6.0
+
+### Minor Changes
+
+- Bump
+
+### Patch Changes
+
+- Updated dependencies
+  - gamba-core-v2@0.4.0
+  - gamba-react-v2@2.0.0
+
+## 0.5.2
+
+### Patch Changes
+
+- Fixed WagerInput issues
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies
+  - gamba-react-v2@1.1.6
+
+## 0.5.0
+
+### Minor Changes
+
+- Improvements
+
+### Patch Changes
+
+- Updated dependencies
+  - gamba-react-v2@1.1.5
+
 ## 0.4.6
 
 ### Patch Changes